requirements.txt
asgiref==3.8.1
Django==5.0.2
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
PyJWT==2.8.0
sqlparse==0.5.1
